agent/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
langchain==0.2.16
langchain-core==0.2.38
langchain-anthropic==0.1.23
langchain-text-splitters==0.2.4

